+++ src/utils/enums/affine-server-preset-id.enum.ts
@@ -1,5 +1,4 @@
 export enum AffineServerPresetId {
-  MASTERYHUB = 'masteryhub',
   CLOUD = 'cloud',
   CUSTOM = 'custom',
 }